// packages/core/src/core/types.ts
export enum BodyId {
  Sun = 'Sun',
  Moon = 'Moon',
  Mercury = 'Mercury',
  Venus = 'Venus',
  Mars = 'Mars',
  Jupiter = 'Jupiter',
  Saturn = 'Saturn',
  Uranus = 'Uranus',
  Neptune = 'Neptune',
  Pluto = 'Pluto',
  MeanNode = 'MeanNode',
  TrueNode = 'TrueNode',
  SouthNode = 'SouthNode',
  LilithMean = 'LilithMean',
  LilithTrue = 'LilithTrue',
  Vertex = 'Vertex',
  AntiVertex = 'AntiVertex',
  ParsFortunae = 'ParsFortunae',
  Chiron = 'Chiron'
}
// ...
export interface CelestialPosition {
  id: BodyId;
  name: string;
  degree: number;    // 0-30 degrees within sign
  longitude: number; // 0-360 degrees (Ecliptic Longitude)
  latitude: number;  // ecliptic latitude
  declination: number; 
  distance: number;  // AU
  speed: number;     // degrees per day
  isRetrograde: boolean;
  house: number;     // 1-12
  sign: string;      // Aries, Taurus, etc.
  emoji: string;
}

export interface HouseCusp {
  house: number; // 1-12
  degree: number;    // 0-30 degrees within sign
  longitude: number; // 0-360 degrees
  sign: string;
  emoji: string;
}

export interface ChartAngles {
  Asc: number;
  MC: number;
  Dsc: number;
  IC: number;
  Vertex?: number;
}

export enum AspectType {
  Conjunction = 'Conjunction',
  Opposition = 'Opposition',
  Trine = 'Trine',
  Square = 'Square',
  Sextile = 'Sextile'
}

export interface Aspect {
  body1: CelestialPosition;
  body2: CelestialPosition;
  type: AspectType;
  angle: number; 
  orb: number;   
  isApplying: boolean; 
}